+++ gax/src/main/java/com/google/api/gax/rpc/InvalidArgumentException.java
@@ -43,4 +43,9 @@ public class InvalidArgumentException extends ApiException {
       String message, Throwable cause, StatusCode statusCode, boolean retryable) {
     super(message, cause, statusCode, retryable);
   }
+
+  public InvalidArgumentException(
+      Throwable cause, StatusCode statusCode, boolean retryable, ErrorDetails errorDetails) {
+    super(cause, statusCode, retryable, errorDetails);
+  }
 }