+++ gax/src/main/java/com/google/api/gax/rpc/UnauthenticatedException.java
@@ -42,4 +42,9 @@ public class UnauthenticatedException extends ApiException {
       String message, Throwable cause, StatusCode statusCode, boolean retryable) {
     super(message, cause, statusCode, retryable);
   }
+
+  public UnauthenticatedException(
+      Throwable cause, StatusCode statusCode, boolean retryable, ErrorDetails errorDetails) {
+    super(cause, statusCode, retryable, errorDetails);
+  }
 }